refactor: update imports to use mock.server module
