+++ src/app/community/_components/CommentList.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from "react";
 import { Comment } from "../comments/Comment";
-import { getComments } from "@/api/server/community/mock";
+import { getComments } from "@/api/server/community/mock.server";
 import { CommentProps } from "@/types/community";
 
 export default function CommentList({
