+++ src/app/community/posts/[id]/edit/page.tsx
@@ -12,7 +12,7 @@ import {
   getCommunity,
   openGraph,
   updateCommunity,
-} from "@/api/server/community/mock";
+} from "@/api/server/community/mock.server";
 import { useRouter } from "next/navigation";
 import OpenGraphPreview from "@/app/_components/OpenGraphLinkReview";
